backend/requirements.txt
Flask>=2.2
flask_cors
requests
pandas
openai
python-dotenv
Flask-Limiter